requirements.txt
pillow==5.4.1
matplotlib==3.0.3
numpy==1.16.2
pandas==0.24.2
seaborn==0.9.0
statsmodels==0.9.0
jupyter==1.0.0
tensorflow-gpu==1.13.1
dominate==2.3.1
scikit-image==0.14.2